Integrate wire type regressions into existing input tests
